+++ lib/plugins/index.js
@@ -16,7 +16,7 @@ module.exports = [
   require('./rollback/index.js'),
   require('./slstats/slstats.js'),
   require('./plugin/plugin.js'),
-  require('./plugin/install/install.js'),
+  require('./plugin/install.js'),
   require('./plugin/uninstall/uninstall.js'),
   require('./plugin/list/list.js'),
   require('./plugin/search/search.js'),
