+++ lib/plugins/plugin/install.js
@@ -5,10 +5,10 @@ const childProcess = BbPromise.promisifyAll(require('child_process'));
 const fse = require('fs-extra');
 const path = require('path');
 const _ = require('lodash');
-const yamlAstParser = require('../../../utils/yamlAstParser');
-const fileExists = require('../../../utils/fs/fileExists');
-const pluginUtils = require('../lib/utils');
-const npmCommandDeferred = require('../../../utils/npm-command-deferred');
+const yamlAstParser = require('../../utils/yamlAstParser');
+const fileExists = require('../../utils/fs/fileExists');
+const pluginUtils = require('./lib/utils');
+const npmCommandDeferred = require('../../utils/npm-command-deferred');
 
 class PluginInstall {
   constructor(serverless, options) {
