+++ test/unit/lib/plugins/plugin/lib/utils.test.js
@@ -7,7 +7,7 @@ const path = require('path');
 const fse = require('fs-extra');
 const proxyquire = require('proxyquire');
 const chalk = require('chalk');
-const PluginInstall = require('../../../../../../lib/plugins/plugin/install/install');
+const PluginInstall = require('../../../../../../lib/plugins/plugin/install');
 const Serverless = require('../../../../../../lib/Serverless');
 const CLI = require('../../../../../../lib/classes/CLI');
 const { expect } = require('chai');
